superskills/slide-designer/requirements.txt
jinja2>=3.1.0
pyyaml>=6.0
markdown>=3.4.0
